add api sink c++
